refactor: implement logic for reading and writing legacy repo version
for use in migrations
