+++ pkg/repo/fs.go
@@ -2,7 +2,7 @@ package repo
 
 import (
 	"context"
-	"encoding/json"
+	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -44,6 +44,7 @@ func NewFS(params FsRepoParams) (*FsRepo, error) {
 	}, nil
 }
 
+// Path returns the filesystem path to of the repo directory.
 func (fsr *FsRepo) Path() (string, error) {
 	if exists, err := fsr.Exists(); err != nil {
 		return "", err
@@ -53,6 +54,7 @@ func (fsr *FsRepo) Path() (string, error) {
 	return fsr.path, nil
 }
 
+// Exists returns true if the repo exists and is valid, false otherwise.
 func (fsr *FsRepo) Exists() (bool, error) {
 	// check if the path is present
 	if _, err := os.Stat(fsr.path); os.IsNotExist(err) {
@@ -60,21 +62,21 @@ func (fsr *FsRepo) Exists() (bool, error) {
 	} else if err != nil {
 		return false, err
 	}
+	var version int
 	// check if the system metadata file is present
-	metaPath := fsr.join(SystemMetadataFile)
-	if _, err := os.Stat(metaPath); os.IsNotExist(err) {
-		if found, err := fsr.legacyVersionExists(); err != nil {
-			return false, nil
-		} else if found {
-			return true, nil
+	if _, err := os.Stat(fsr.join(SystemMetadataFile)); os.IsNotExist(err) {
+		// if it's not present check if this repo is pre version 4, which requires a migration
+		version, err = fsr.readLegacyVersion()
+		if err != nil {
+			return false, err
 		}
-		return false, nil
 	} else if err != nil {
 		return false, err
-	}
-	version, err := fsr.readVersion()
-	if err != nil {
-		return false, err
+	} else {
+		version, err = fsr.readVersion()
+		if err != nil {
+			return false, err
+		}
 	}
 	if !IsValidVersion(version) {
 		return false, NewUnknownRepoVersionError(version)
@@ -82,53 +84,16 @@ func (fsr *FsRepo) Exists() (bool, error) {
 	return true, nil
 }
 
-func (fsr *FsRepo) legacyVersionExists() (bool, error) {
-	versionPath := fsr.join(VersionFile)
-	if _, err := os.Stat(versionPath); os.IsNotExist(err) {
-		return false, nil
-	} else if err != nil {
-		return false, err
-	}
-	versionBytes, err := os.ReadFile(versionPath)
-	if err != nil {
-		return false, err
-	}
-	var version Version
-	if err := json.Unmarshal(versionBytes, &version); err != nil {
-		return false, err
-	}
-	if !IsValidVersion(version.Version) {
-		return false, NewUnknownRepoVersionError(version.Version)
-	}
-	return true, nil
-}
-
-func (fsr *FsRepo) readLegacyVersion() (int, error) {
-	versionPath := fsr.join(VersionFile)
-	versionBytes, err := os.ReadFile(versionPath)
-	if err != nil {
-		return -1, err
-	}
-	var version Version
-	if err := json.Unmarshal(versionBytes, &version); err != nil {
-		return -1, err
-	}
-	if !IsValidVersion(version.Version) {
-		return -1, NewUnknownRepoVersionError(version.Version)
-	}
-	return version.Version, nil
-}
-
 // Version returns the version of the repo.
 func (fsr *FsRepo) Version() (int, error) {
-	return fsr.readVersion()
+	maybeVersion, err := fsr.readLegacyVersion()
+	if errors.Is(os.ErrNotExist, err) {
+		return fsr.readVersion()
+	}
+	return maybeVersion, nil
 }
 
-// join joins path elements with fsr.path
-func (fsr *FsRepo) join(paths ...string) string {
-	return filepath.Join(append([]string{fsr.path}, paths...)...)
-}
-
+// Init initializes a new repo, returning an error if the repo already exists.
 func (fsr *FsRepo) Init(c config.ReadWriter) error {
 	if exists, err := fsr.Exists(); err != nil {
 		return err
@@ -168,6 +133,7 @@ func (fsr *FsRepo) Init(c config.ReadWriter) error {
 	return fsr.WriteVersion(Version4)
 }
 
+// Open opens an existing repo, returning an error if the repo is uninitialized.
 func (fsr *FsRepo) Open(c config.ReadWriter) error {
 	// if the repo does not exist we cannot open it.
 	if exists, err := fsr.Exists(); err != nil {
@@ -273,7 +239,7 @@ func (fsr *FsRepo) WriteRunInfo(ctx context.Context, summaryShellVariablesString
 	*/
 }
 
-// modifies the config to include keys for accessing repo paths
+// EnsureRepoPathsConfigured modifies the config to include keys for accessing repo paths
 func (fsr *FsRepo) EnsureRepoPathsConfigured(c config.ReadWriter) {
 	c.SetIfAbsent(types.AuthTokensPath, fsr.join(config.TokensPath))
 	c.SetIfAbsent(types.UserKeyPath, fsr.join(config.UserPrivateKeyFileName))
@@ -287,4 +253,9 @@ func (fsr *FsRepo) EnsureRepoPathsConfigured(c config.ReadWriter) {
 
 	c.SetIfAbsent(types.NodeRequesterJobStorePath, fsr.join(config.OrchestratorStorePath, "jobs.db"))
 	c.SetIfAbsent(types.NodeComputeExecutionStorePath, fsr.join(config.ComputeStorePath, "executions.db"))
+}
+
+// join joins path elements with fsr.path
+func (fsr *FsRepo) join(paths ...string) string {
+	return filepath.Join(append([]string{fsr.path}, paths...)...)
 }
